save vcmt and some refactoring
